Fix a flurznrc bug reading the SLOTE parameter
Fix a minor bug in FLURZnrc where it expected the wrong number of inputs
when reading in the SLOTE parameter from the egsinp file. This did not
result in any actual issues, but an error message was printed to the
screen about the user providing the wrong number of inputs.